// src/game/enums.ts
export enum ProvinceFeature {
	Pasture = "pasture",
	Fishery = "fishery",
	GameLand = "gameLand",
	SilverOre = "silverOre",
	FertileSoil = "fertileSoil",
	MarbleDeposits = "marbleDeposits",
	MineralDeposits = "mineralDeposits",
	BrineDeposits = "brineDeposits",
}

export enum Product {
	Honey = "honey",
	Meat = "meat",
	Dyes = "dyes",
	Hides = "hides",
	Ink = "ink",
	Parchment = "parchment",
	Tackling = "tackling",
	Salt = "salt",
	Statues = "statues",
	Silver = "silver",
	Wool = "wool",
	Columns = "columns",
	Horses = "horses",
	Clothes = "clothes",
	Leather = "leather",
	IllustratedBooks = "illustratedBooks",
	Wax = "wax",
	Linen = "linen",
	Wine = "wine",
}

export enum BuildingId {
	Admiralty = "admiralty",
	Armoury = "armoury",
	AxeSmith = "axeSmith",
	Bakery = "bakery",
	BallistaTowers = "ballistaTowers",
	BeeYard = "beeYard",
	Bulwark = "bulwark",
	Butcher = "butcher",
	Catapults = "catapults",
	Cathedral = "cathedral",
	CattleFarm = "cattleFarm",
	Cauldrons = "cauldrons",
	ChainMailWorkshop = "chainMailWorkshop",
	Church = "church",
	CoastGuard = "coastGuard",
	CornerTowers = "cornerTowers",
	Docks = "docks",
	DrumTowers = "drumTowers",
	DyesWorkshop = "dyesWorkshop",
	Fishmongery = "fishmongery",
	Fletcher = "fletcher",
	GateTowers = "gateTowers",
	Granary = "granary",
	HalberdMastersmithy = "halberdMastersmithy",
	Harbour = "harbour",
	HempField = "hempField",
	Hostel = "hostel",
	HuntersHuts = "huntersHuts",
	InkMaker = "inkMaker",
	Inn = "inn",
	Library = "library",
	Market = "market",
	MerchantGuild = "merchantGuild",
	Moat = "moat",
	Palisade = "palisade",
	ParchmentMaker = "parchmentMaker",
	Pickler = "pickler",
	PlateArmourer = "plateArmourer",
	Quarry = "quarry",
	RiggersStore = "riggersStore",
	SaltMine = "saltMine",
	ScaleArmourer = "scaleArmourer",
	ScribesOffice = "scribesOffice",
	SculptorsGuild = "sculptorsGuild",
	SheepFarm = "sheepFarm",
	SiegeWorkshop = "siegeWorkshop",
	SilverMine = "silverMine",
	SpearMaker = "spearMaker",
	SpinningMill = "spinningMill",
	Stable = "stable",
	StoneWall = "stoneWall",
	Stonemason = "stonemason",
	Stud = "stud",
	SwordMastersmithy = "swordMastersmithy",
	Swordsmith = "swordsmith",
	Tailor = "tailor",
	Tannery = "tannery",
	TaxCollectorsOffice = "taxCollectorsOffice",
	ToolSmithy = "toolSmithy",
	TownWatchHouse = "townWatchHouse",
	TrainingGrounds = "trainingGrounds",
	University = "university",
	Vineyard = "vineyard",
	WaxMaker = "waxMaker",
	WeavingMill = "weavingMill",
	Winery = "winery",
}

export enum GlobalResource {
	Gold = "gold",
	Piety = "piety",
	Books = "books",
}

export enum LocalResource {
	Workers = "workers",
	Food = "food",
}